extract callPhone and sendEmail from Tool.js
